+++ Sources/KeyboardKit/_Keyboard/Keyboard+Services.swift
@@ -68,9 +68,7 @@ public extension Keyboard {
         }
 
         /// The dictation service to use.
-        public lazy var dictationService: DictationService = .disabled(
-            context: state.dictationContext
-        )
+        public lazy var dictationService: DictationService = .disabled
 
         /// The feedback service to use.
         public lazy var feedbackService: KeyboardFeedbackService = .standard